Change to avm_evt.c to fix issue seen on HP Proliant rack mount.
Ticket #449

This change only affects the operation of HP Proliant rack mount
servers.  The problem corrected is that shutting down an HP Proliant
rack mount server that is running OpenSAF, causes the server to
reboot instead of halt.  The fix is to manually advance the AVM
state instead of calling avm_active_inactive(), which by-passes
code that is attempting to communicate with the server as it is
already shutting down.  This interference is what causes the server
to reboot instead of shutdown.  These changes have been unit tested
and correct the issue that is described in ticket #449.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>