Use a different HPL MDS timeout value for HP Proliant.
No ticket.

This is a minor enhancement to use a different timeout
value for HPL MDS messages that are targeted for HP's
Proliant rack mount servers.  The default timeout value
is too short for this particular platform, so it is
changed to a higher value if the HP PROLIANT platform
is detected.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>