MDS Linear TIPC address scheme support
Ticket #473

A new TIPC address scheme is added to MDS. The new scheme uses TIPC addresses
in the form 1.1.1, 1.1.2, 1.1.3 etc. This is achieved by not using the subslot
ID in generation of the TIPC node part of the TIPC network address. The change
is fully backwards compatible by default and will require changes to CFLAGS and
script.conf to enable the new addressing scheme.