immsv: Removed unnecessary delay in starting IMMND sync (#470)
#470

In immnd_proc.c in the periodic check that runs once a second in idle
state, the check for sync is now done every time, instead of every fifth
time. In addition, when sync request is detected, the period is immediately
reduced to the non-idle period of 0.1 sec. Any sync request should now cause
a sync to be started in aproximately one second (two on UML).