dtsv: optionally manage archives of dtsv files
Ticket #205 & #292

When configured with DTSV_MAX_ARCHIVES in script.conf, archive management is
enabled. When a cluster restart of OpenSAF is done, current set of files will
be archived (in the background) and thus moved from the set of active files.

Max DTSV_MAX_ARCHIVES archives will be stored in the ${NCS_LOG_PATH}/archives
directory and the oldest archive is removed.

An archive is a gzipped tar ball.