Publish all HPI events - instead of filtering some of them.
Ticket #482

HISv is suppose to receive HPI events and publish them to the EDSv channel.
However, as it stands now HISv heavily filters HPI events based on which entity
generated the event. On some platforms this level of filtering causes key
events, such as an enclosure fan failure, to get dropped instead of being
published to EDSv. This patch adds a function to publish these "other" events
instead of dropping them.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>