Fix for issues on HP c-Class and Proliant Servers
Ticket #454

This patch fixes issues found on HP c-Class and Proliant rack mount
servers only.  The construction of entity-paths in FM for sub-slots
is adjusted, and in AVM, the issues of shutting down or locking an
HP Proliant server from the CLI is cleaned up - so that is works
properly without side-effects on the OpenSAF controller-side or on
the node that is told to shutdown.

Signed-off-by: Michael Bishop <michael.bishop@hp.com>