immsv: Local unpack and pre-check of fevs before forward (#2747 & #1871)
This patch adds a framework for pre-checking fevs packed messages, in the
local IMMND. The idea is to catch corrupt messages in the local IMMND,
before they are forwarded to IMMD and broadcast to all IMMNDs.
Thus preventing all IMMNDs from crashing on the same corrupt and broadcast
message.

The checks done in this framwork added in this patch are quite modest and
more checks can and should be added later. A 100% solution for #2747 can
never be reached so I will close #2747 with the push of these changesets
under review.

The pre-check also takes care of suporting local idempotency checks, fixing
the re-opened enhancement #1871. Finally, the pre-check also removes the
issue of local write checks being controlled from the imma library. The
library code is still the same and the checks are currently still done in
the local IMMND, but the value for the pre-check flag is now verified for
each message type. In a later cleanup I will remove the usage of the
write-check flag from the imma library support of fevs.