smfd: handles single node system node reboot upgrade (#2904)
If the system only contain one AMF node, a rolling upgrade where a bundle requires reboot will be treated as a cluster reboot.
If the class OpenSafSmfMisc is defined in the system, the rollback operation will be rejected if the campaign contain a cluster reboot procedure.
Smfd works also without the OpenSafSmfMisc class but will not reject rollback of campaigns containing cluster reboot procedures.