immsv: Consumed search/accessor is de-allocated by server (#3007)
When a search_next is about to return the termination condition
'SA_AIS_ERR_NOT_EXIST' from the local immnd server, it dealocates the search
operation. This is for both saImmOmSearchNext and saImmOmAccessorGet. The
latter actually uses searchNext internally. if the iteration was consumed
before the finalize then saImmOmSearchFinalize will be a no-op in the server.
A search that is finalized before being fully consumed works as before,
i.e. will deallocate the remaining server structure. The library code
for saImmOmAccessorGet actually uses searchInitialize and searchNext,
as before.