Logsv: Fix logd process abort (#3022)
In the amf-callbacks of Logsv, Logsv calls immutil functions,
e.g immutil_saImmOiImplementerSet. The default behaviour for
these utility functions is that error situations results in
process abort, i.e IMM aborts the process.
The correct behaviour should be that Logsv always returns the
amf-callback, also when an immutil call is unsuccessful. It is
then up to AMF to act. To accomplish this, Logsv resets the
'errorsAreFatal' setting in appropriate places.