amfnd: fix admin comp restart to terminate [#7]
Since a common routine was used for both error handling and admin comp restart,
cleanup was done also in the admin case. This was because the comp got marked
"FAILED" also in the admin restart case.

A new function comp_restart_initiate() is added. This function does not set the
FAILED state on a component. Instead the error handling logic does it before
calling the new function. From admin restart context, FAILED state is of course
not set resulting in either term callback or term CLC CLI execution.