IMM: IMMD verifies file configuration for IMMNDs [#17]
IMMNDs send their file configuration parameters to IMMD.
IMMD verifies that all IMMNDs agree on configuration.
The first IMMND at an SC that joins will be taken as the "correct"
configuration by the active IMMD. If IMMND at the other SC does not
match the configuration it will not be allowed to join (load or sync).
Payload IMMNDs must only agree that PBE *is* configured or not.
If 2PBE is configured for IMMD, then PBE must be configured for IMMNDs.