IMM: 2PBE patch-2 (dumping) [#21]
This patch contains the 2PBE dumping mechanism, needed to support 2PBE.
A PBE process is started by the IMMND at each SC, not just the IMMND coord.

The PBE colocated with the IMMND coord, called the primary PBE, is still the
coordinator for transaction commits (for CCBs and PRT operations and
class-create/deletes). The primary PBE (sometimes called the A-side PBE)
works very much in the same way as the regular single PBE does in 1-PBE.
The PBE colocated with the SC resident but non-coord IMMND is called
the slave PBE (sometimes called the B-side PBE).

With 2PBE, *both* PBEs must be available for the imm to be persistent-
writable. If one or boths PBEs are unavailable (or unresponsive) then
persistent writes (ccbs and PRT operations) will fail.

In 2PBE, a restarted PBE will more often need to regenerate the sqlite
file. On the other hand, regeneration of the sqlite file should be faster
in 2PBE than in regular 1PBE because the file is typically placed on a
local file system.

A subsequent patch will provide a mechhanism for allowing 1-safe-2PBE.
This will allow the imm to open up for persistent writes when only one
of the two PBEs are available. THis will only be allowed when and during
the absence of an SC. As soon as the other SC rejoins the IMM has to
re enter the non-persistent-writable state.