IMM: 2PBE patch-6 (PBE slave can re-attach when empty CCBs exist) [#21]
When a 2PE system is up and running with ccbs being generated, if one SC is
rebooted, then after the SC has synced imm-ram, the slave pbe typically has
trouble in being allowed to generate its imm.db.xxxxx file. It keeps getting
rejected due to active ccbs. There realy should not be any active ccbs allowed
here because the sync of the returned SC would only have started when there
are no active ccbs and once sync is finished the imm should still not be
persistent writable. THe only problem hewre is that empty CCBs are allowed
to be created. Thus the condiftion for allowing the slave to generate its
imm.db.xxxx file needs to be relaxed to allow empty CCBs.