mds: use TIPC segmentation/reassembly [#654]
TIPC supports sending message with a maximum length of 66000 bytes. MDS supports
sending even bigger messages, upto ~46Mb. MDS has used an internal fragmentation
of just 1400 bytes (MDTM_NORMAL_MSG_FRAG_SIZE) despite that the receive buffer
is 8000 bytes (MDS_DIRECT_BUF_MAXSIZE)

This patch changes MDS to use TIPC segmentation/reassembly. This gives benefits
such as secure transmission (TIPC does retransmission). TIPC send congestion is
reduced since TIPC counts messages not bytes. Less processing done in user
space in the sending opensaf or client process.

Since the receive buffer of older MDS cores is limited to 8000 bytes we have be
sure not to send larger messages than that to older cores. Otherwise we have
problems with in-service upgradeability.

Each MDS TIPC port is bound to a functional address - a port name. The port name
consists of 32 bit type and 32 bit instance. The 4 msb of the instance word
contains the so called "archword" field. The msb indicates if this MDS core is
32 or 64 bit. The remaining 3 bits is now used as a version field. Version 0 is
used by older versions of MDS, version 1 is now used to indicate the capability
for longer fragmentation.

This information is part of the published port name thus a subscriber will
get them using the TIPC discovery service. These bits are ignored by old MDS
cores, so it can receive messages from a new MDS core. When sending a message
this bit is used to understand if the peer uses TIPC segmentation/reassembly
(or not) and adapt the limit accordingly.

The possibility to set mds_arch using configure has been removed.

The change is in-service compliant and can be upgraded into a system.